// app/types.tsx
// types.ts
export interface RecordData {
    _id: any;
    year: string;
    month: string;
    date_received: string;
    log_time: string;
    serial_no: string;
    from_whom_received: string;
    date_of_letter: string;
    letter_ref_no: string;
    received_by: string;
    type_of_letter: string;
    subject: string;
  }
  